pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shelf-analyzer"
version = "0.1.0"
description = "FastAPI-based shelf analysis using YOLO and LLMs"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "ultralytics==8.0.220",
    "opencv-python==4.8.1.78",
    "Pillow==10.1.0",
    "numpy==1.25.2",
    "openai==1.3.7",
    "google-cloud-aiplatform==1.38.1",
    "anthropic==0.7.8",
    "pandas==2.1.4",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "aiofiles==23.2.0",
    "setuptools>=68.0.0",
    "wheel>=0.41.0"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.1"
]
